requirements.txt
numpy
opencv-python
matplotlib
scikit-image
tqdm
scipy
google-colab

# Replace with the appropriate cupy package for your CUDA version
cupy-cuda112
